+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## dev
+
+### Added
+
+- support multi-sample VCF files according to issue #35
+
 ## v1.0.0 - [24th October 2023]
 
 Initial release of variantinterpretation pipeline, created with the [nf-core](https://nf-co.re/) template.